requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
colorthief==0.2.1
Pillow==11.1.0
httpx==0.28.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
